updated add travel expenses details correctly

+++ app/src/main/java/com/tekskills/er_tekskills/data/model/AddPurposeMeetingResponse.kt
@@ -74,7 +74,7 @@ data class UserCordinatesResponse(
     @SerializedName("tenantId")
     val tenantId: Int,
     @SerializedName("totalDistance")
-    val totalDistance: Int,
+    val totalDistance: Double,
     @SerializedName("updatedAt")
     val updatedAt: String,
     @SerializedName("updatedBy")
